+++ PhoneShop/src/main/java/com/product/phoneshop/exception/ResourceNotFoundException.java
@@ -9,9 +9,9 @@ import org.springframework.http.HttpStatus;
 public class ResourceNotFoundException extends ServiceException{
 
     private String resourceName;
-    private Integer resourceId;
+    private Long resourceId;
 
-    public ResourceNotFoundException( String resourceName , Integer resourceId) {
+    public ResourceNotFoundException( String resourceName , Long resourceId) {
         super(HttpStatus.NOT_FOUND,String.format("%s not found for id= %d",resourceName ,resourceId));
     }
 }
