+++ PhoneShop/src/main/java/com/product/phoneshop/repository/ColorRepository.java
@@ -3,6 +3,6 @@ package com.product.phoneshop.repository;
 import com.product.phoneshop.model.Color;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ColorRepository extends JpaRepository<Color, Integer> {
+public interface ColorRepository extends JpaRepository<Color, Long> {
 
 }
