+++ PhoneShop/src/main/java/com/product/phoneshop/repository/ModelRepository.java
@@ -4,6 +4,6 @@ import com.product.phoneshop.model.Model;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface ModelRepository extends JpaRepository<Model, Integer>, JpaSpecificationExecutor<Model> {
+public interface ModelRepository extends JpaRepository<Model, Long>, JpaSpecificationExecutor<Model> {
 
 }
