+++ PhoneShop/src/main/java/com/product/phoneshop/service/ModelService.java
@@ -7,7 +7,7 @@ import java.util.Map;
 public interface ModelService {
 
     Model save(Model dto) ;
-    Model getById(Integer id);
+    Model getById(Long id);
     //List<Model> getModels(Map<String,String> params);   //we want to use dynamic query by using Map
     Page<Model> getModels(Map<String, String> params);
