+++ PhoneShop/src/main/java/com/product/phoneshop/service/impl/ModelServiceImpl.java
@@ -29,7 +29,7 @@ public class ModelServiceImpl implements ModelService {
 
     //Get By ID
     @Override
-    public Model getById(Integer id)  {
+    public Model getById(Long id)  {
         return modelRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Model" , id));
     }
@@ -42,13 +42,13 @@ public class ModelServiceImpl implements ModelService {
         ModelFilter modelFilter = new ModelFilter();
 
         if(params.containsKey("modelId")) {
-            modelFilter.setModelId(MapUtils.getInteger(params, "modelId"));
+            modelFilter.setModelId(MapUtils.getLong(params, "modelId"));
         }
         if(params.containsKey("modelName")) {
             modelFilter.setModelName(MapUtils.getString(params, "modelName"));
         }
         if(params.containsKey("brandId")) {
-            modelFilter.setBrandId(MapUtils.getInteger(params, "brandId"));
+            modelFilter.setBrandId(MapUtils.getLong(params, "brandId"));
         }
         if(params.containsKey("brandName")) {
             modelFilter.setBrandName(MapUtils.getString(params, "brandName"));
